[PiP] Maintain origin for the initiator of active PiP session.
According to the spec[1], the use agent has a list of origins for
initiators of active PiP sessions. This CL makes
PictureInPictureWindowController maintain the origin and add an
interface in PictureInPictureWindowManager for querying the list
of origins for initiators of active PiP sessions.

[1] https://www.w3.org/TR/picture-in-picture/#defines

Bug: none
Change-Id: I3559e3d77d112948fc52a1f12d4ddeb2330a7764
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4853229
Reviewed-by: Fr <beaufort.francois@gmail.com>
Commit-Queue: Wei4 Wang <wei4.wang@intel.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197011}