Mitigate recovery failure due to unstable network during Gaia login
RART fetching could fail due to unstable network during initial Gaia
frame loading, and it will not re-fetch when the network is back on.
This would result in an error screen during recovery due to missing
RAPT. A proper fix would require refactoring the network handling logic
in `GaiaScreenHandler` and re-fetch the RART when the network is back
on.

As a temporary mitigation, we'll show a reauth needed message when the
RAPT is not present, and ask the user to go through reauth again, at
which point the network will likely become stable.

Bug: b/300218067
Change-Id: I398550cbca5fe71d4ac43a3e53d073e9811b501c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4861734
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Yunke Zhou <yunkez@google.com>
Cr-Commit-Position: refs/heads/main@{#1197020}