[QuickStart] Add quickstart component
This CL creates and uses a quick-start-button component which will be used instead of the oobe-icon-button with the benefit of having the quick start button more self contained and reusable.

This will help in future tasks for viewing an "enable bluetooth" dialog when the user choses to use quick start but their bluetooth is disabled (see b/283965454).

Bug: b/298972956
Change-Id: I95fc8772e46fe0c93e749e9c81818317e54860d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4831551
Commit-Queue: Renato Silva <rrsilva@google.com>
Reviewed-by: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/main@{#1197054}