Simplify header copy in SpdyStream::SaveResponseHeaders
SaveResponseHeaders() copied the headers by iterating over them. Make it
simpler by assigning a clone of the headers instead. This also may be
slightly faster as it avoids the check if the header already exists.

Also use the contains() method on the headers for a tiny improvement in
readability.

Change-Id: I7ffb4dca7ffadfe27ef037346c2ec8d2617ba496
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4865526
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197060}