[OOBE] - Remove exapnd button in oobe apps list component
Remove expand button for an app element which truncated description.
There is a 80 character limit on this field, so we will not exceed two
lines in any case.
Updated debug.js to include one item with a short description and one
with a long one.

Old 1:
https://screenshot.googleplex.com/6yWcMssFfA7kEBr
Old 2:
https://screenshot.googleplex.com/AVQqYqpVDWJGKzD
New:
https://screenshot.googleplex.com/AiwwZJQXummcZ4y

Bug: b:296609871
Change-Id: I8830a83796d711efc85b272e84f7e4a15559e0e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4865266
Commit-Queue: Bohdan Tyshchenko <bohdanty@google.com>
Reviewed-by: Danila Kuzmin <dkuzmin@google.com>
Cr-Commit-Position: refs/heads/main@{#1197072}