[Android][CredMan] Introduce component scaffold for No-Passkeys sheet
The component consists only of
* Coordinator (creating the static message),
* Mediator (showing the static message and handling dismiss), and
* SheetContent (replacing the View)
since there is no property that changes after creation of the message.

The mediator is passed as a delegate for actions triggered on the
sheet content object which is ineffective while no UI to interact with
exists. Tests of delegate methods are added as soon as there is UI since
they are functionally covered by the module tests.

Bug: 1481495
Change-Id: I75191f10f3773e0ccd2c4588af1a9af14f409d8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4866532
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Friedrich Horschig <fhorschig@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1197077}