[iOS] Add test for reopening History Sync when previous one is shown
It's possible to tap on the history sync promo action button to show a
new history sync screen, while the dismiss animation is being played
for the previous History Sync screen, causing previously a DCHECK to be
triggered in the first coordinator's dealloc, since it has never been
stopped.

This CL adds a test to verify that DCHECK is not triggered in this case.

Fixed: 1480475
Change-Id: I982cd16c13e3d50394a7d04d6bd18e7338dd2e01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4866123
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197081}