[profiles] Plumb CoreAccountInfo to the ProfileNameResolver
The profile name resolver needs the email of the account to use as
default name when fetching the user name fails.
Previously, it was trying to get the email from the account ID, but
this is difficult: in some cases the email is in the primary account
info, and sometimes it is attached to an non-primary account.

To avoid adding complex logic in the profile name resolver to find
the email, the email is passed in the constructor (as part of the
CoreAccountInfo).

Fixed: 1481902
Change-Id: Ic020e0024499d028573373c858790367a13f1586
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4859355
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: Gabriel Oliveira <gabolvr@google.com>
Cr-Commit-Position: refs/heads/main@{#1197133}