Introduce a feature flag to do early RFH swaps for navigation transitions
This CL introduces a new kind of early RenderFrameHost swap, intended
to support navigation transitions for back/forward navigations.
Namely, this unloads the old document early, without waiting for the
navigation to a new document to commit, such that it is safe to show a
preview of the new document while waiting for its navigation to
complete.

This is done behind a new feature that's disabled by default,
EarlyDocumentSwapForBackForwardTransitions.  It is also added to
chrome://flags/#early-document-swap-for-back-forward-transitions.

This swap occurs later than the current early swap for non-live
(initial or crashed) frames, after DidStartNavigation is dispatched to
observers and after WillStartRequest throttles finish processing. This
is intentional, as it avoids performing the early swap if the
navigation fails prior to this point, and it avoids disrupting
existing implementations of DidStartNavigation and WillStartRequest.
Because WillStartRequest() may defer the navigation, this early swap
may be processed on a different task than the one where
NavigationRequest is created and the speculative RenderFrameHost and
its SiteInstance is computed.  Eventually, the goal is to move the
existing early swap cases to this new call site as well in
https://crbug.com/1467011.

The swap is only used for cross-document, browser-initiated
back/forward navigations (reloads are excluded) in outermost main
frames.  A navigation must use a speculative RenderFrameHost for the
swap to be possible.  Navigations that are the result of activations,
namely back-forward cache restores and prerender activations, also do
not perform this early swap, as they follow a different code path and
the early swap is actually not needed in those cases.

The swap currently only works for cross-process navigations; there is
a known issue where same-process navigations (e.g., when
RenderDocument is used) fail some blink widget visibility checks,
which will be fixed in a later CL.

Change-Id: I78b58f4115a6ba8ddb1ddd5e1bbecf1f94b528a0
Bug: 1480129
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4847998
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197193}