Stop geolocation watch updates upon permission revocation
Previously, once a site obtained a geolocation watch (either through an already existing grant or through a new permission prompt flow), it kept receiving location updates in the registered watch callback within the same session (i.e. for as long as the user didn't close the tab or performed a navigation). This remained the case, even after the grant was revoked. This CL fixes this behaviour.

Bug: 1462313
Change-Id: I3151beb1a05bcf7169e952e908b67ae9dbcf7105
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4853341
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197210}