dlp: Expect download item having an invalid GURL
There are cases in which DownloadItem::GetURL() returns an invalid GURL.
In this case we cannot perform a dlp check. If there is actually a dlp
restriction, keeping the default path, could lead to dlp blocking the
actual save operation. In this case the user could not successfully save
the file. So we prompt for the save dialog so the user can chose a path.
This might be unexpected as there is no real reason for prompting. But
this only happens to users using dlp.

BUG=b:300605501

Change-Id: I05dd59d2ece389b6d11f22e31e020eca86951756
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4868374
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Daniel Brinkers <brinky@google.com>
Cr-Commit-Position: refs/heads/main@{#1197229}