[iOS] Add egtests for skipped optional History Sync Opt-In
The history sync promo should not be shown too frequently to the users.
In the flow for which history sync is optional, the opt-in screen should
be skipped if any user declined it in the past 2 weeks, or twice in a
row.

This CL add few EGtests for the prompt triggered in Recent Tabs (where
the History Sync Opt-In step is mandatory), and in the settings (where
the step is optional). It also extract few matchers for the History Sync
UI to facilitate testing.

Bug: 1450867
Change-Id: I2426c552b7fa1385d07039ed05343d8f1d6e4e3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4852096
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197252}