Create network request to retrieve all plus addresses.
This will be used so that the local map will be kept in sync with the
remote enterprise-specified plus address server. This request is not
made yet, but will be occur whenever the PersistedRepeatingTimer
is triggered. That change will come in a later CL.

This CL also does a small refactor to consolidate some of the parsing
logic into its own class and updates the affected tests.

Bug: 1467623 b/299668285 b/299670826 b/300123836
Change-Id: I7859ee362fe2fa31f18caec28105c518a1157554
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4858864
Commit-Queue: Kirubel Aklilu <kaklilu@chromium.org>
Reviewed-by: Side YILMAZ <sideyilmaz@chromium.org>
Reviewed-by: Matt Reichhoff <mreichhoff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197258}