[Nearby Presence] Add PublicCredentialType mojom conversion.
To facilitate passing the ::nearby::presence::PublicCredentialType enum
as a parameter to mojo interfaces, this CL adds a conversion to a new
::ash::nearby::presence::proto::mojom::PublicCredentialType enum.

PublicCredentialType is used to denote whether a PublicCredential
is considered remote or local. It is used as a parameter when
saving and querying for Nearby Presence credentials over the
`nearby_presence_credential_storage.mojom` interface.

A full design for Nearby Presence Credential
Storage can be found at go/cros-nearby-presence-credential-storage

Test: Unit tests
Bug: b:287334195
Change-Id: I203165eefea5b9e1b0f1560273afccd075be7b9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4807324
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Juliet Lévesque <julietlevesque@google.com>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Mark Polner <polner@google.com>
Cr-Commit-Position: refs/heads/main@{#1197290}