[dPWA] Fix heap UAF in WebAppUninstallDialogView
There is an UAF triggered in WebAppUninstallDialogView where uninstall
with a ScopedKeepAlive wipes out the dialog view, leading to a UAF in
the checkbox_ stored in the dialog view. There might be other
concerns as well about clearing web app site data breaking, since that
is called after the checkbox_ is accessed.

This CL fixes that by:
1. Storing the information about checkbox before triggering uninstall,
and ensuring that Uninstall is the last call happening.
2. Moving the logic to clear web app site data to the caller. This
allows the site clearing logic to be tied to the ScopedKeepAlive
callback, allowing site clearing to happen before the keep alive
destructs, thus preventing any possible UAFs.

Bug: 1481063
Change-Id: Ic96b29cc45e54b00a7ae4faf85f727b36d869d02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4862702
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197331}