Create a disable_crashes parameter for the kMetricsReportingFeature.
We want to support uploading metrics and not crashes.
This currently only supports Windows.

The logic here is pretty odd and I don't really love it, but I don't wan to a big restructure in the middle of adding a new feature and it's not clearly wrong. The subtle thing here is there is 2 things used on the crash side, the registry key (which the logic here is setting the Feature being on) and also the consent parameter (which currently includes both consent,the feature, and other ideas).

We decided to just include this parameter for the registry key part and not include it in the consent parameter since it isn't related to consent. This is still functional since both are checked.

We tested end to end on a local windows build.

Change-Id: I0f1d429a240e07e73cd9c934846f2cc675dc8db2

Bug: 1482819
Change-Id: I0f1d429a240e07e73cd9c934846f2cc675dc8db2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4855367
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Commit-Queue: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197341}