media/gpu/v4l2/test: Remove comment
The comment is technically correct, forward probabilities update
uAPI changed in 5.17. This feature has never been implemented
in either the platform decoder nor the chrome stack decoder.
Therefore the comment is confusing as it refers to a feature
that has not and is not on the roadmap and is not helpful as
the comment does not specify where and how this control should
be used.

BUG=none

Change-Id: I799f4e9b1140f66ce2942022468b76e3cc5ecce4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4868323
Reviewed-by: Steve Cho <stevecho@chromium.org>
Commit-Queue: Steve Cho <stevecho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197367}