[dPWA] Create tests for force OS unregistration policy for web apps
This CL creates some unit tests that verify some use-cases and ensures
that things work fine with force OS unregistration applied in the
WebAppSettings policy. This is being done separately so as not to
bloat the actual CL in the chain, which can be backported all the
way to M116.

Bug: 1482803, b/300429265
Change-Id: I6d81a7a45e83d76b6e2cf7dd8478226658e68a41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4865597
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197368}