Rename GetMaximizedOrCenteredWindowType
Rename GetMaximizedOrCenteredWindowType to GetWindowTypeOnMaximizable
inside WindowState and LockWindowState. Updated the corresponding
comments as well.
For the one inside `WindowState`, which is used in tablet mode and the
window will be centered if it is not maximizable. But for the one
inside LockWindowState, which is used in the lock screen. The window
will take the full work area even it is in normal state, no centered
related logic at all.

Bug: b/300164495
Change-Id: Idb2e3e89333d973a66b4e04c111a8d3f70bc69e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4863940
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Min Chen <minch@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197373}