Add WebP bpp (bits per pixel) histograms
Add a family of 23 Blink.DecodedImage.WebPDensity.Count.* histograms to
record the bits-per-pixel values of WebP images for various image sizes
(where size is the area = width * height).

- 9 histograms for image sizes 0.1 MP ~ 0.9 MP, in increments of 0.1 MP
- 13 histograms for image sizes 1 MP - 13 MP, in increments of 1 MP
- 1 histogram for image sizes > 13 MP

To avoid recording outliers in these histograms, only lossy, still (not
animated) images without an alpha plane are recorded in these
histograms.

The histograms are updated when the images are successfully decoded.

Change the comparison
    demux_state_ <= WEBP_DEMUX_PARSING_HEADER
to the equivalent comparison
    demux_state_ < WEBP_DEMUX_PARSED_HEADER
and change
    DCHECK_GT(demux_state_, WEBP_DEMUX_PARSING_HEADER)
to
    DCHECK_GE(demux_state_, WEBP_DEMUX_PARSED_HEADER)
so that demux_state_ is compared with WEBP_DEMUX_PARSED_HEADER
consistently.

Bug: 1456182
Change-Id: Ia876ad9845f7aee112d7a9ec319edaace96fc965
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4828711
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Wan-Teh Chang <wtc@google.com>
Cr-Commit-Position: refs/heads/main@{#1197466}