[VC] Disable StopAllScreenShare.
There is currently a known crash caused by this.

(1) When a user is using google doc/slice, they can directly present from that page to a google Meet meeting.
(2) This will add two stopCallbacks for the same device to the display_media_stop_callbacks_.
(3) If the user stop the screen sharing from the webui, these two stop_callbacks will be removed correctly.
(4) However if the user uses the StopAllScreenShare button in the vc tray, these two stopCallbacks will both be called, and calling the second one will cause a crash.

This will influence any caller of StopMediaCapturing which contains
both ChromeOS (Video Conferencing) and Android (Digital Wellbeing).

This cl is only a temporary fix that not recording any stopcallbacks,
so no callback will be called when clicking on, therefore no crash.

We want to keep the change for easier cherry pick. We'll land a proper
fix later.

BUG=chromium:1479984

TEST='Tested on Dut manually"

Change-Id: Ia4525c3db904fbb6404fe63f3e696a1dcea7dbcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4865213
Commit-Queue: Charles Zhao <charleszhao@chromium.org>
Reviewed-by: Elad Alon <eladalon@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197526}