Explicitly set features in RetainEthernetIPAddr tests
The *NetworkPolicyApplication* browser_tests that care about the
default "Recommended" handling of Ethernt IP address fields (i.e. that
depend on the enablement of DisablePolicyEthernetRecommendedWorkaround)
started behaving differently on different builders with CL:4866161 due
to testing/variations/fieldtrial_testing_config.json not being applied
on all builders.

This makes these tests explicitly set the feature to the expected value.
While we're at it, we're returning to one RetainEthernetIPAddr test with
the feature enabled and one with the feature disabled for clarity.

Bug: b:271988376, 1483397
Change-Id: I0ef026282f16d7ad0c91cefd7508e0a811596779
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4864951
Reviewed-by: Anqing Zhao <anqing@chromium.org>
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197648}