[Files app] Ducks: Add Selectors to the store
Introduces the concept of Selectors to the store.

Selectors are essentially functions that allow for a parts of the
store's state to be combined to create a value that can be read
at any moment in time, subscribed to for changes, or used directly
in lit components in way that allows them to automatically
re-render when selectors update.

Selectors can be combined to create new selectors and are
implemented efficiently through a DAG.

Selectors only emit new values when the fail the shallow
equality check with their previous state, freeing components
and other subscribers from the potentially tedius and repetitive
job of checking for equality when raw store updates are received.

Additionally, the SelectorEmitter class ensures that even if
multiple parents of a given selector emit, their common child
selector only emits once, instead of once for every parent that
has changes. It also guarantees that a selector's new value is
only ever calculated when at least one of its parents emits a
new value.

This CL adds unit tests in order to showcase how selectors can
be used in Files app.

Bug: b:297808212
Test: browser_tests FileManagerJsTest.Selector
Change-Id: I12f68fec6d15d5a59725654cc94d7a48ead3f3b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4861070
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Marcello Salomao <msalomao@google.com>
Cr-Commit-Position: refs/heads/main@{#1197672}