[iOS] Migrate usage of ConsentLevel::kSync in omnibox autocomplete
Background: ConsentLevel::kSync is going away, and all users need
to be migrated to something else - see crbug.com/1447010.

Before this CL, AutocompleteProviderClientImpl::IsAuthenticated()
checked for ConsentLevel::kSync, which seemed wrong (independent of
the removal of kSync): This method is meant to return whether
there's a signed-in user, not whether they have Sync enabled (per
the comment in autocomplete_provider_client.h, and also judging
from its call sites).
So this CL updates it to use ConsentLevel::kSignin instead (behind
a killswitch, just in case).

Note that this is also more similar to the implementation on other
platforms, which checks for any accounts in the cookie jar (i.e. is
also independent of Sync).

Bug: 1462286, 1483272
Change-Id: Ie5ddf0065a2206e2f3157ac425948085a00aab6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4866343
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Auto-Submit: Marc Treib <treib@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197719}